massive change induced by the use of "size_t" in yarp
svn path=/trunk/iCub/; revision=14421

+++ src/libraries/iKin/tutorials/genericChainController/main.cpp
@@ -61,7 +61,7 @@ protected:
         if (vect.length()!=b.size())
             vect.resize(b.size());
 
-        for (int i=0; i<vect.length(); i++)
+        for (size_t i=0; i<vect.length(); i++)
             vect[i]=b.get(i).asDouble();
 
         mutex.post();
